finished the compile errors

+++ exercises/exercise11.cpp
@@ -11,7 +11,7 @@ int main() {
 
   const size_t loopUpperLimit = 5;
 
-  for (int index = 0; index < loopUpperLimit; ++index) {
+  for (size_t index = 0; index < loopUpperLimit; ++index) {
     cout << "index is " << index << endl;
   }
